backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
aiofiles==23.2.1

# LLM & AI
openai==1.12.0
chromadb==0.4.24

# Document Processing
pymupdf==1.23.26
python-docx==1.1.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.2.1
loguru==0.7.2

# HTTP Client
httpx==0.27.0
